requirements.txt
openai>=1.2.3
requests>=2.31.0
python-dotenv>=1.0.0
pyttsx3>=2.90
streamlit
streamlit-chat
BeautifulSoup